Refactoring the controllers and services .  Add verfication to middleware.
