+++ projects/authentication_management/zainab1AT/Server.js
@@ -1,7 +1,7 @@
 import express from "express";
 import dotenv from "dotenv";
-import authRoutes from "./authRoutes.js";
-import userController from "./userController.js";
+import userRoutes from "./routes/userRoutes.js";
+import authRoutes from "./routes/authRoutes.js";
 
 const app = express();
 
@@ -11,8 +11,8 @@ const PORT = process.env.PORT;
 
 app.use(express.json());
 
-app.use("/api/auth", userController);
-app.use("/user", authRoutes);
+app.use("/api/auth", authRoutes);
+app.use("/api", userRoutes);
 
 app.listen(PORT, () => {
   console.log(`Server is running on port ${PORT}`);
